backend/requirements.txt
annotated-types==0.7.0
attrs==25.1.0
Authlib==1.4.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0